Update LICENSE
The reason for the change in license to be more strict is to reinforce the fact this exists to hold code and have it run once. Any mod choosing to reuse the code internally instead of simply making this a dependency, asking for potential additions/options to the code where it'd be easier for me to do so instead, is choosing to make your game slower if you happen to use both, running the code twice, when it only needed to run once.